♻️ refactor: 주식 차트 그리기 함수 utils 폴더로 이동 #60
